Reorganized the model code to be based on layers.
