--- meteor-app/imports/api/models/comment/index.ts
@@ -1,4 +1,0 @@
-export * from "./comment-id";
-export * from "./comment-mutations";
-export * from "./comment-queries";
-export * from "./comment";
