--- meteor-app/imports/api/models/company/company-id.ts
@@ -1,11 +1,0 @@
-import { PgId, Branded } from "imports/api/models";
-
-export type CompanyId = Branded<PgId, "CompanyId">;
-
-export function companyIdToString(id: CompanyId): string {
-	return String(id);
-}
-
-export function stringToCompanyId(id: string): CompanyId {
-	return Number(id) as CompanyId;
-}
