--- meteor-app/imports/api/models/company/index.ts
@@ -1,4 +1,0 @@
-export * from "./company-id";
-export * from "./company-mutations";
-export * from "./company-queries";
-export * from "./company";
