--- meteor-app/imports/api/models/job-ad/index.ts
@@ -1,4 +1,0 @@
-export * from "./job-ad-id";
-export * from "./job-ad-mutations";
-export * from "./job-ad-queries";
-export * from "./job-ad";
