--- meteor-app/imports/api/models/job-ad/job-ad.ts
@@ -1,16 +1,0 @@
-import { JobAdId, CompanyId } from "imports/api/models";
-
-export type JobAd = {
-	jobAdId: JobAdId;
-
-	companyName: string | null;
-	companyId: CompanyId;
-
-	jobTitle: string;
-	pesosPerHour: string;
-	contractType: string;
-	jobDescription: string;
-	responsibilities: string;
-	qualifications: string;
-	dateAdded: Date;
-};
