--- meteor-app/imports/api/models/location/index.ts
@@ -1,2 +1,0 @@
-export * from "./location-queries";
-export * from "./location";
