+++ meteor-app/imports/api/models/mutations/index.ts
@@ -1,0 +1,4 @@
+export * from "./company";
+export * from "./review";
+export * from "./salary";
+export * from "./vote";
