+++ meteor-app/imports/api/models/mutations/salary.ts
@@ -6,7 +6,7 @@ import {
 
 import CreateSalaryInput from "imports/lib/inputs/salary";
 
-import { SalaryId, Salary, UserPId } from "imports/api/models";
+import { SalaryId, UserPId } from "imports/api/models";
 
 export async function createSalary(
 	input: CreateSalaryInput,
@@ -77,15 +77,4 @@ export async function createSalary(
 	};
 
 	return execTransactionRW(transaction);
-}
-
-export async function editSalary(
-	id: SalaryId,
-	salaryChanges: unknown
-): Promise<Salary> {
-	throw new Error("Not implemented yet");
-}
-
-export async function deleteSalary(id: SalaryId): Promise<Salary> {
-	throw new Error("Not implemented yet");
 }
