+++ meteor-app/imports/api/models/queries/index.ts
@@ -1,0 +1,8 @@
+export * from "./comment";
+export * from "./company";
+export * from "./job-ad";
+export * from "./review";
+export * from "./salary";
+export * from "./user";
+export * from "./vote";
+export * from "./location";
