+++ meteor-app/imports/api/models/queries/salary.ts
