+++ meteor-app/imports/api/models/queries/vote.ts
