--- meteor-app/imports/api/models/review/index.ts
@@ -1,4 +1,0 @@
-export * from "./review-id";
-export * from "./review-mutations";
-export * from "./review-queries";
-export * from "./review";
