--- meteor-app/imports/api/models/review/review-id.ts
@@ -1,11 +1,0 @@
-import { PgId, Branded } from "imports/api/models";
-
-export type ReviewId = Branded<PgId, "ReviewId">;
-
-export function reviewIdToString(id: ReviewId): string {
-	return String(id);
-}
-
-export function stringToReviewId(id: string): ReviewId {
-	return Number(id) as ReviewId;
-}
