--- meteor-app/imports/api/models/salary/index.ts
@@ -1,4 +1,0 @@
-export * from "./salary-id";
-export * from "./salary-mutations";
-export * from "./salary-queries";
-export * from "./salary";
