--- meteor-app/imports/api/models/user/index.ts
@@ -1,4 +1,0 @@
-export * from "./user-id";
-export * from "./user-mutations";
-export * from "./user-queries";
-export * from "./user";
