--- meteor-app/imports/api/models/user/user.ts
@@ -1,9 +1,0 @@
-import { UserMId, CompanyId } from "imports/api/models";
-
-export type User = {
-	_id: UserMId;
-	username: string;
-	createdAt: Date;
-	role: "worker" | "company-unverified" | "company";
-	companyId: CompanyId | null;
-};
