--- meteor-app/imports/api/models/vote/index.ts
@@ -1,4 +1,0 @@
-export * from "./vote-id";
-export * from "./vote-mutations";
-export * from "./vote-queries";
-export * from "./vote";
